Move out sensor controlled spin into separate file
